Split meca out from base_plotting.py
