+++ pygmt/figure/__init__.py
@@ -65,6 +65,9 @@ class Figure(BasePlotting):
         self._preview_dir = TemporaryDirectory(prefix=self._name + "-preview-")
         self._activate_figure()
 
+    # GMT Supplementary modules
+    from .meca import meca
+
     def __del__(self):
         # Clean up the temporary directory that stores the previews
         if hasattr(self, "_preview_dir"):
